tuotteen poistonapin refaktorointi typeskriptiksi, nappi listauksiin

+++ client/src/components/DeleteProduct.tsx
@@ -3,8 +3,13 @@ import { Button } from 'react-bootstrap'
 import productService from '../services/products'
 import { useStore  } from '../store'
 import { useHistory } from 'react-router-dom'
+import { Product } from '../types'
 
-const DeleteProduct = ({ product }) => {
+type Props = {
+    product: Product
+}
+
+const DeleteProduct: React.FC<Props> = ({ product }) => {
   const productCreatorId = product.user
   const { user, setNotification, products, setProducts } = useStore()
   const history = useHistory()
@@ -13,15 +18,17 @@ const DeleteProduct = ({ product }) => {
     return (null)
   }
 
-  const handleClick = async (event) => {
+  const handleClick: React.MouseEventHandler<HTMLElement> = async (event) => {
     event.preventDefault()
-    try {
-      await productService.remove(product.id)
-      setProducts(products.filter(p => p.id !== product.id))
-      history.push('/products')
-      setNotification(`Tuote ${product.name} poistettu onnistuneesti`, 'success')
-    } catch(error) {
-      setNotification('Tuotteen poistaminen ei onnistunut', 'error')
+    if (window.confirm(`Poista tuote ${product.name}`)) {
+      try {
+        await productService.remove(product.id)
+        setProducts(products.filter(p => p.id !== product.id))
+        history.push('/products')
+        setNotification(`Tuote ${product.name} poistettu onnistuneesti`, 'success')
+      } catch(error) {
+        setNotification('Tuotteen poistaminen ei onnistunut', 'error')
+      }
     }
   }
